requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
pydantic==2.5.3
aiogram==3.4.1
psutil==5.9.8
GPUtil==1.4.0
python-multipart==0.0.20 
alembic==1.12.0
python-jose==3.4.0
passlib==1.7.4
asyncpg==0.30.0